# public/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Music API",
    "version": "1.0.0",
    "description": "A RESTful API for a music streaming service, built with Node.js, Express, MongoDB and Cloudinary",
    "contact": {
      "name": "Music API Support"
    }
  },
  "servers": [
    {
      "url": "https://music-api-qxuc.vercel.app",
      "description": "Production server"
    },
    {
      "url": "http://localhost:5000",
      "description": "Development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Enter JWT token"
      }
    },
    "schemas": {
...
      "Song": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "description": "Song ID"
          },
          "title": {
            "type": "string",
            "description": "Song title"
          },
          "artist": {
            "type": "string",
            "description": "Artist ID"
          },
          "album": {
            "type": "string",
            "description": "Album ID"
          },
          "duration": {
            "type": "number",
            "description": "Duration in seconds"
          },
          "audioUrl": {
            "type": "string",
            "format": "uri",
            "description": "URL to audio file"
          },
          "coverImage": {
            "type": "string",
            "format": "uri",
            "description": "URL to cover image"
          },
          "releaseDate": {
            "type": "string",
            "format": "date-time",
            "description": "Release date"
          },
          "genre": {
            "type": "string",
            "description": "Song genre"
          },
          "plays": {
            "type": "number",
            "description": "Number of plays",
            "default": 0
          },
          "likes": {
            "type": "number",
            "description": "Number of likes",
            "default": 0
          },
          "isExplicit": {
            "type": "boolean",
            "description": "Whether song contains explicit content",
            "default": false
          },
          "featuredArtists": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Array of featured artist IDs"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "Error": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "description": "Error message"
          },
          "status": {
            "type": "string",
            "description": "Error status"
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/api/songs/create": {
      "post": {
        "summary": "Create a new song (Admin only)",
        "description": "Create a new song with audio file and optional cover image",
        "tags": [
          "Songs"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": [
                  "title",
                  "artist",
                  "duration",
                  "audio"
                ],
                "properties": {
                  "title": {
                    "type": "string",
                    "description": "Song title"
                  },
                  "artist": {
                    "type": "string",
                    "description": "Artist ID"
                  },
                  "album": {
                    "type": "string",
                    "description": "Album ID"
                  },
                  "duration": {
                    "type": "number",
                    "description": "Duration in seconds"
                  },
                  "audio": {
                    "type": "string",
                    "format": "binary",
                    "description": "Audio file"
                  },
                  "coverImage": {
                    "type": "string",
                    "format": "binary",
                    "description": "Cover image file"
                  },
                  "genre": {
                    "type": "string"
                  },
                  "isExplicit": {
                    "type": "boolean"
                  },
                  "featuredArtists": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    },
                    "description": "Array of featured artist IDs"
                  },
                  "releaseDate": {
                    "type": "string",
                    "format": "date-time"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Song created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Song"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - Admin access required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/songs/{id}": {
      "get": {
        "summary": "Get song by ID",
        "description": "Retrieve detailed information about a specific song",
        "tags": [
          "Songs"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Song ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Song retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Song"
                }
              }
            }
          },
          "404": {
            "description": "Song not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Update song details (Admin only)",
        "description": "Update an existing song's information",
        "tags": [
          "Songs"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Song ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": {
                    "type": "string"
                  },
                  "artist": {
                    "type": "string"
                  },
                  "album": {
                    "type": "string"
                  },
                  "genre": {
                    "type": "string"
                  },
                  "isExplicit": {
                    "type": "boolean"
                  },
                  "featuredArtists": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Song updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Song"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - Admin access required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Song not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete a song (Admin only)",
        "description": "Delete a song from the database",
        "tags": [
          "Songs"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Song ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Song deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - Admin access required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Song not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/songs/{id}/add-album": {
      "post": {
        "summary": "Add album to song (Admin only)",
        "description": "Associate an album with a song",
        "tags": [
          "Songs"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Song ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "albumId"
                ],
                "properties": {
                  "albumId": {
                    "type": "string",
                    "description": "Album ID to add"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Album added to song successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Song"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - Admin access required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Song or album not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/songs/{id}/add-artist": {
      "post": {
        "summary": "Add artist to song (Admin only)",
        "description": "Associate an artist with a song",
        "tags": [
          "Songs"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Song ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "artistId"
                ],
                "properties": {
                  "artistId": {
                    "type": "string",
                    "description": "Artist ID to add"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Artist added to song successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Song"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - Admin access required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Song or artist not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}